+++ homeassistant/block_async_io.py
@@ -5,7 +5,7 @@ import importlib
 import sys
 import time
 
-from .util.async_ import protect_loop
+from .util.loop import protect_loop
 
 
 def enable() -> None:
