+++ homeassistant/util/async_.py
@@ -5,15 +5,9 @@ from __future__ import annotations
 from asyncio import AbstractEventLoop, Future, Semaphore, Task, gather, get_running_loop
 from collections.abc import Awaitable, Callable, Coroutine
 import concurrent.futures
-from contextlib import suppress
-import functools
-import linecache
 import logging
-import sys
 import threading
 from typing import Any, ParamSpec, TypeVar, TypeVarTuple
-
-from homeassistant.exceptions import HomeAssistantError
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -92,145 +86,6 @@ def run_callback_threadsafe(
         raise RuntimeError("The event loop is in the process of shutting down.")
 
     return future
-
-
-def check_loop(
-    func: Callable[..., Any],
-    strict: bool = True,
-    strict_core: bool = False,
-    advise_msg: str | None = None,
-    **mapped_args: Any,
-) -> None:
-    """Warn if called inside the event loop. Raise if `strict` is True.
-
-    The default advisory message is 'Use `await hass.async_add_executor_job()'
-    Set `advise_msg` to an alternate message if the solution differs.
-    """
-    if (
-        func.__name__ == "import_module"
-        and (args := mapped_args.get("args"))
-        and args[0] in sys.modules
-    ):
-        # If the module is already imported, we can ignore it.
-        return
-
-    try:
-        get_running_loop()
-        in_loop = True
-    except RuntimeError:
-        in_loop = False
-
-    if not in_loop:
-        return
-
-    # Import only after we know we are running in the event loop
-    # so threads do not have to pay the late import cost.
-    # pylint: disable=import-outside-toplevel
-    from homeassistant.core import HomeAssistant, async_get_hass
-    from homeassistant.helpers.frame import (
-        MissingIntegrationFrame,
-        get_current_frame,
-        get_integration_frame,
-    )
-    from homeassistant.loader import async_suggest_report_issue
-
-    found_frame = None
-
-    if func.__name__ == "sleep":
-        #
-        # Avoid extracting the stack unless we need to since it
-        # will have to access the linecache which can do blocking
-        # I/O and we are trying to avoid blocking calls.
-        #
-        # frame[1] is us
-        # frame[2] is protected_loop_func
-        # frame[3] is the offender
-        with suppress(ValueError):
-            offender_frame = get_current_frame(3)
-            if offender_frame.f_code.co_filename.endswith("pydevd.py"):
-                return
-
-    offender_frame = get_current_frame(2)
-    offender_filename = offender_frame.f_code.co_filename
-    offender_lineno = offender_frame.f_lineno
-    offender_line = (
-        linecache.getline(offender_filename, offender_lineno) or "?"
-    ).strip()
-
-    try:
-        integration_frame = get_integration_frame()
-    except MissingIntegrationFrame:
-        # Did not source from integration? Hard error.
-        if not strict_core:
-            _LOGGER.warning(
-                "Detected blocking call to %s with args %s in %s, "
-                "line %s: %s inside the event loop",
-                func.__name__,
-                mapped_args.get("args"),
-                offender_filename,
-                offender_lineno,
-                offender_line,
-            )
-            return
-
-        if found_frame is None:
-            raise RuntimeError(  # noqa: TRY200
-                f"Detected blocking call to {func.__name__} inside the event loop "
-                f"in {offender_filename}, line {offender_lineno}: {offender_line}."
-                f"{advise_msg or 'Use `await hass.async_add_executor_job()`'}; "
-                "This is causing stability issues. Please create a bug report at "
-                f"https://github.com/home-assistant/core/issues?q=is%3Aopen+is%3Aissue"
-            )
-
-    hass: HomeAssistant | None = None
-    with suppress(HomeAssistantError):
-        hass = async_get_hass()
-    report_issue = async_suggest_report_issue(
-        hass,
-        integration_domain=integration_frame.integration,
-        module=integration_frame.module,
-    )
-
-    _LOGGER.warning(
-        (
-            "Detected blocking call to %s inside the event loop by %sintegration '%s' "
-            "at %s, line %s: %s, (offender: %s, line %s: %s) please %s"
-        ),
-        func.__name__,
-        "custom " if integration_frame.custom_integration else "",
-        integration_frame.integration,
-        integration_frame.relative_filename,
-        integration_frame.line_number,
-        integration_frame.line,
-        offender_filename,
-        offender_lineno,
-        offender_line,
-        report_issue,
-    )
-
-    if strict:
-        raise RuntimeError(
-            "Blocking calls must be done in the executor or a separate thread;"
-            f" {advise_msg or 'Use `await hass.async_add_executor_job()`'}; at"
-            f" {integration_frame.relative_filename}, line {integration_frame.line_number}:"
-            f" {integration_frame.line} "
-            f"(offender: {offender_filename}, line {offender_lineno}: {offender_line})"
-        )
-
-
-def protect_loop(
-    func: Callable[_P, _R], strict: bool = True, strict_core: bool = True
-) -> Callable[_P, _R]:
-    """Protect function from running in event loop."""
-
-    @functools.wraps(func)
-    def protected_loop_func(*args: _P.args, **kwargs: _P.kwargs) -> _R:
-        check_loop(
-            func, strict=strict, strict_core=strict_core, args=args, kwargs=kwargs
-        )
-        return func(*args, **kwargs)
-
-    return protected_loop_func
 
 
 async def gather_with_limited_concurrency(
